Revert "Creation/Modification of bank info only for Account Payment group"
The previous commit breaks test in accunt module for invoice creation.

Also partner managers should be allowed to change bank numbers. If finer
grained authority is needed, it should be done in a way that does not
break existing code. Anyway when registering customers/partners it
would be very restrictive if a person from accounting department would
need to be involved.

This reverts commit b1eac3f6e4e619fa962895852265a3c261ba56df.